Bug fix float type difference np and torch

+++ src/data_loader/dataset.py
@@ -71,7 +71,7 @@ class HyperspectralFruitDataset(Dataset):
         if isinstance(img_data, np.ndarray):
             img_data = img_data.astype(np.float32)
         tensor = self.transform(img_data)
-        
+        tensor = tensor.float()
         return tensor, torch.tensor(sample['label'], dtype=torch.long)
